test: LottoServiceTestProvider 분리
